working working very hard

+++ board/src/main/java/ch/bfh/akka/botrace/board/model/game/items/weapon/Bow.java
@@ -3,7 +3,7 @@ package ch.bfh.akka.botrace.board.model.game.items.weapon;
 public class Bow extends RangedWeapon {
 
 
-    public Bow(String name, int weight, int attackValue) {
+    public Bow(String name, int weight, double attackValue) {
         super(name, weight, attackValue);
         this.setItemDescription("Ranged weapon" +
                 "\nAttack Strength: " + attackValue);
