+++ board/src/main/java/ch/bfh/akka/botrace/board/model/game/items/weapon/Club.java
@@ -2,7 +2,7 @@ package ch.bfh.akka.botrace.board.model.game.items.weapon;
 
 
 public class Club extends MeleeWeapon {
-    public Club(String name, int weight, int attackValue, int defenceValue) {
+    public Club(String name, int weight, double attackValue, double defenceValue) {
         super(name, weight, attackValue, defenceValue);
         this.setItemDescription("Melee weapon" +
                 "\nAttack Strength: " + attackValue +
