+++ board/src/main/java/ch/bfh/akka/botrace/board/model/game/items/weapon/RangedWeapon.java
@@ -2,7 +2,7 @@ package ch.bfh.akka.botrace.board.model.game.items.weapon;
 
 public abstract class RangedWeapon extends Weapon{
 
-    public RangedWeapon(String name, int weight, int attackValue) {
+    public RangedWeapon(String name, int weight, double attackValue) {
         super(name, weight, attackValue);
     }
 }
