+++ board/src/main/java/ch/bfh/akka/botrace/board/model/game/items/weapon/Sword.java
@@ -1,7 +1,7 @@
 package ch.bfh.akka.botrace.board.model.game.items.weapon;
 
 public class Sword extends MeleeWeapon {
-    public Sword(String name, int weight, int attackValue, int defenceValue) {
+    public Sword(String name, int weight, double attackValue, double defenceValue) {
         super(name, weight, attackValue, defenceValue);
         this.setItemDescription("Melee weapon" +
                 "\nAttack Strength: " + attackValue +
