+++ board/src/main/java/ch/bfh/akka/botrace/board/model/game/items/weapon/ThrowingKnive.java
@@ -1,7 +1,7 @@
 package ch.bfh.akka.botrace.board.model.game.items.weapon;
 
 public class ThrowingKnive extends RangedWeapon {
-    public ThrowingKnive(String name, int weight, int attackValue) {
+    public ThrowingKnive(String name, int weight, double attackValue) {
         super(name, weight, attackValue);
         this.setItemDescription("Ranged weapon" +
                 "\nAttack Strength: " + attackValue);
